requirements.txt
fastapi~=0.75.2
uvicorn==0.17.6
wget==3.2
numpy==1.22.4
pydantic==1.9.1
torchvision== 0.15.2
torch==2.0.1
Pillow==9.1.1
locust==2.15.1

